complete till stock copy no

+++ src/main/java/com/raja/lib/invt/repository/StockCopyNoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.raja.lib.invt.model.StockCopyNo;
 
-public interface StockCopyNoRepository extends JpaRepository<StockCopyNo, Integer> {
+public interface StockCopyNoRepository extends JpaRepository<StockCopyNo, Long> {
 
 }
